+++ weather-widget/.gitignore
@@ -1,3 +1,4 @@
 test.txt
 test.py
 venv/*
+*/__pycache__
